requirements.txt
streamlit
pandas
python-dotenv
google-generativeai
dateparser
